order_executor/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
etcd3==0.12.0